requirements.txt
flask>=2.3.0
ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
scipy>=1.10.0
requests>=2.31.0
werkzeug>=2.3.0
timm>=0.9.0
pillow-heif>=1.0.0
